Add 'history' parameter to infiniteStream
The history parameter allows users to set an upper bound on
infiniteStream's memory usage, as normally the emitted values are
remembered and then printed to the console.

Note that the regression test snapshots are updated because the stream
is now printing additional information. I assume the error messages are
not part of fast-check's semver guarantees; otherwise this will have to
be reverted.