✨ Add 'history' parameter to infiniteStream (#6107)
**Description**
The new `history` parameter allows users to set an upper bound on
infiniteStream's memory usage by giving up visibility into the items it
emitted. In the future, it would be possible to enrich this parameter to
create a bounded memory window at the beginning and/or end of the
stream, but for now the option is just a boolean to toggle history on
and off.

To still present some useful information, the stream now prints the
count of the number of emitted items. The count shows even when the
history is present, to save the user from manually counting. I assume
the exact string representation of the arbitrary is not part of
fast-check's semver guarantees and so this would be a minor bump. Let me
know if I'm mistaken.

Fixes #6106.

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR includes bumps details, please run `pnpm run bump` and flag
the impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: ...🐛✨
- [x] Impacts: ... should be ok.

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR -->
<!-- * ✨ Introduce new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->

<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change -->
<!-- * Generated values: Can your change impact any of the existing
generators in terms of generated values, if so which ones? when? -->
<!-- * Shrink values: Can your change impact any of the existing
generators in terms of shrink values, if so which ones? when? -->
<!-- * Performance: Can it require some typings changes on user side?
Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->

---------

Co-authored-by: Nicolas DUBIEN <github@dubien.org>