📝 Preserve links on Sponsors for the website (#6169)
**Description**

<!-- Please provide a short description and potentially linked issues
justifying the need for this PR -->

In the past, we used to import the sponsors by relying on an `img` tag.
Unfortunately it dropped all the links.

We switched to an `object` tag with a self-hosted svg file that we pull
from GitHub at build time.

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [ ] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [ ] My PR references one of several related issues (if any)
- [ ] New features or breaking changes must come with an associated
Issue or Discussion
- [ ] My PR does not add any new dependency without an associated Issue
or Discussion
- [ ] My PR includes bumps details, please run `pnpm run bump` and flag
the impacts properly
- [ ] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [ ] Category: ...
- [ ] Impacts: ...

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR -->
<!-- * ✨ Introduce new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->

<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change -->
<!-- * Generated values: Can your change impact any of the existing
generators in terms of generated values, if so which ones? when? -->
<!-- * Shrink values: Can your change impact any of the existing
generators in terms of shrink values, if so which ones? when? -->
<!-- * Performance: Can it require some typings changes on user side?
Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->