📝 Fix example in quick start guide (#6287)
**Description**

This PR fixes an example in the quick start tutorial. Currently the
example reads:


```
test('should keep an already sorted array sorted', () => {
  expect(sortNumbersAscending([1, 2, 3])).toEqual([1, 2, 3]);
});
test('should sort a randomly ordered array in ascending order', () => {
  expect(sortNumbersAscending([3, 1, 2])).toEqual([1, 2, 3]);
});
test('should sort a descending ordered array in ascending order', () => {
  expect(sortNumbersAscending([3, 1, 2])).toEqual([1, 2, 3]);
});

```

Where the third example test description indicates the input values
should be descending order array i.e. `[3, 2, 1]` but currently it shows
`[3, 1, 2]`, which is a duplicate test case from the one above it. This
PR updates so the test correctly reflect the description.

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [ ] My PR references one of several related issues (if any)
- [ ] New features or breaking changes must come with an associated
Issue or Discussion
- [ ] My PR does not add any new dependency without an associated Issue
or Discussion
- [ ] My PR includes bumps details, please run `pnpm run bump` and flag
the impacts properly
- [ ] My PR adds relevant tests and they would have failed without my PR
(when applicable)


**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: 
   * 📝 Add or update documentation 
- [ ] Impacts: ...




<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change -->
<!-- * Generated values: Can your change impact any of the existing
generators in terms of generated values, if so which ones? when? -->
<!-- * Shrink values: Can your change impact any of the existing
generators in terms of shrink values, if so which ones? when? -->
<!-- * Performance: Can it require some typings changes on user side?
Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->