✨ Add `circular` option to `fc.letrec` (#6040)
**Description**

Closes #6039

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [ ] My PR includes bumps details, please run `pnpm run bump` and flag
the impacts properly
- [ ] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: ✨ Introduce new feature
- [x] Impacts: N/A

---

Open questions:
1. Is this really the best place/way to implement this? (see my comments
on #6039)
2. See `// TODO: Do we need to clone here?` in the PR code. If I _do_
need to clone, then what should I be using to do that? Answer: I think
we don't need cloning here as long as we implement shrinking correctly
3. I have not been able to figure out how to implement the logic that
throws if we accidentally don't replace some `placeholderSymbol`. I
tried tracking them in a local set from `nat().map(...)` and then
deleting them later when we replace, and asserting we have zero after,
but this was causing problems, so I'm probably missing something. This
seems like a critical feature in order to ship this
4. Neither `assertProduceValuesShrinkableWithoutContext` nor
`assertShrinkProducesSameValueWithoutInitialContext` pass for `circular:
true`. Does that make sense? Should they pass? If so, then what am I
doing wrong?

---------

Co-authored-by: Nicolas DUBIEN <github@dubien.org>