♻️ Extract logic building lazy arbs
In the context of `letrec`, the `tie` function passed to the caller is responsible to build lazy instances of arbitraries. When first built they cannot generate anything given they have to be given more details on the way to generate the asked instances. They just now the name of the arbitrary they will be responsible to generate but not the how to build it.

In this PR we do several things:

- Re-use the `tie` function exposed to the user internally, so feel the lazy instances once the user returned us its instances
- Move from an Object to a Map for better performance